+++ M295_ILBA24/DTOs/AddressResponseDto.cs
@@ -2,7 +2,7 @@ namespace M295_ILBA24.DTOs;
 
 public class BaseAddressResponseDto
 {
-    public ushort AddressId { get; set; }
+    public int AddressId { get; set; }
 
     public string Address1 { get; set; } = null!;
 
@@ -10,7 +10,7 @@ public class BaseAddressResponseDto
 
     public string District { get; set; } = null!;
 
-    public ushort CityId { get; set; }
+    public int CityId { get; set; }
 
     public string? PostalCode { get; set; }
