+++ M295_ILBA24/DTOs/BaseCountryResponseDto.cs
@@ -2,7 +2,7 @@ namespace M295_ILBA24.DTOs;
 
 public class BaseCountryResponseDto
 {
-    public ushort CountryId { get; set; }
+    public int CountryId { get; set; }
 
     public string Country1 { get; set; } = null!;
 }
