+++ M295_ILBA24/DTOs/CityRequestDto.cs
@@ -3,5 +3,5 @@ namespace M295_ILBA24.DTOs;
 public class CityRequestDto
 {
     public required string City1 { get; set; }
-    public ushort CountryId { get; set; }
+    public int CountryId { get; set; }
 }
