+++ M295_ILBA24/DTOs/CityResponseDto.cs
@@ -2,11 +2,11 @@ namespace M295_ILBA24.DTOs;
 
 public class BaseCityResponseDto
 {
-    public ushort CityId { get; set; }
+    public int CityId { get; set; }
 
     public string City1 { get; set; } = null!;
 
-    public ushort CountryId { get; set; }
+    public int CountryId { get; set; }
     
     public virtual BaseCountryResponseDto Country { get; set; } = null!;
 }
