+++ M295_ILBA24/Entities/Actor.cs
@@ -2,7 +2,7 @@
 
 public partial class Actor
 {
-    public ushort ActorId { get; set; }
+    public int ActorId { get; set; }
 
     public string FirstName { get; set; } = null!;
