+++ M295_ILBA24/Entities/ActorInfo.cs
@@ -5,7 +5,7 @@ namespace M295_ILBA24.Entities;
 
 public partial class ActorInfo
 {
-    public ushort ActorId { get; set; }
+    public int ActorId { get; set; }
 
     public string FirstName { get; set; } = null!;
