+++ M295_ILBA24/Entities/Address.cs
@@ -5,7 +5,7 @@ namespace M295_ILBA24.Entities;
 
 public partial class Address
 {
-    public ushort AddressId { get; set; }
+    public int AddressId { get; set; }
 
     public string Address1 { get; set; } = null!;
 
@@ -13,7 +13,7 @@ public partial class Address
 
     public string District { get; set; } = null!;
 
-    public ushort CityId { get; set; }
+    public int CityId { get; set; }
 
     public string? PostalCode { get; set; }
