+++ M295_ILBA24/Entities/Category.cs
@@ -5,7 +5,7 @@ namespace M295_ILBA24.Entities;
 
 public partial class Category
 {
-    public byte CategoryId { get; set; }
+    public int CategoryId { get; set; }
 
     public string Name { get; set; } = null!;
