+++ M295_ILBA24/Entities/Country.cs
@@ -5,7 +5,7 @@ namespace M295_ILBA24.Entities;
 
 public partial class Country
 {
-    public ushort CountryId { get; set; }
+    public int CountryId { get; set; }
 
     public string Country1 { get; set; } = null!;
