+++ M295_ILBA24/Entities/Customer.cs
@@ -5,9 +5,9 @@ namespace M295_ILBA24.Entities;
 
 public partial class Customer
 {
-    public ushort CustomerId { get; set; }
+    public int CustomerId { get; set; }
 
-    public byte StoreId { get; set; }
+    public int StoreId { get; set; }
 
     public string FirstName { get; set; } = null!;
 
@@ -15,7 +15,7 @@ public partial class Customer
 
     public string? Email { get; set; }
 
-    public ushort AddressId { get; set; }
+    public int AddressId { get; set; }
 
     public bool? Active { get; set; }
