+++ M295_ILBA24/Entities/FilmActor.cs
@@ -5,9 +5,9 @@ namespace M295_ILBA24.Entities;
 
 public partial class FilmActor
 {
-    public ushort ActorId { get; set; }
+    public int ActorId { get; set; }
 
-    public ushort FilmId { get; set; }
+    public int FilmId { get; set; }
 
     public DateTime LastUpdate { get; set; }
