+++ M295_ILBA24/Entities/Inventory.cs
@@ -7,9 +7,9 @@ public partial class Inventory
 {
     public uint InventoryId { get; set; }
 
-    public ushort FilmId { get; set; }
+    public int FilmId { get; set; }
 
-    public byte StoreId { get; set; }
+    public int StoreId { get; set; }
 
     public DateTime LastUpdate { get; set; }
